+++ satella/files.py
@@ -29,18 +29,28 @@ def read_re_sub_and_write(path: str, pattern: tp.Union[re.compile, str],
         f_out.write(data)
 
 
-def find_files(path: str, wildcard: str = r'(.*)') -> tp.Iterator[str]:
+def find_files(path: str, wildcard: str = r'(.*)',
+               prefix_with: tp.Optional[str] = None,
+               scan_subdirectories: bool = True) -> tp.Iterator[str]:
     """
     Look at given path's files and all subdirectories and return an iterator of
     file names (paths included) that conform to given wildcard.
 
     :param path: path to look into.
     :param wildcard: a regular expression to match
+    :param prefix_with: an optional path component to prefix before the filename with os.path.join
+    :param scan_subdirectories: whether to scan subdirectories
     :return: paths with the files. They will be relative paths, relative to path
     """
     for filename in os.listdir(path):
-        if os.path.isdir(os.path.join(path, filename)):
+        if scan_subdirectories and os.path.isdir(os.path.join(path, filename)):
             for file in find_files(os.path.join(path, filename), wildcard):
-                yield os.path.join(filename, file)
+                v = os.path.join(filename, file)
+                if prefix_with is not None:
+                    v = os.path.join(prefix_with, v)
+                yield v
         elif re.match(wildcard, filename):
+            if prefix_with is not None:
+                filename = os.path.join(prefix_with, filename)
             yield filename
+